Updated GitHub "release" action to use jreleaser to create release.
JReleaser will generate a GitHub Release and post the artifacts to the
GitHub Packages repo.  The Release will contain a change list.

Need to wait for JReleaser 1.12 to publish to maven central.  It appears
that Maven Central is in transition to a new API (the publishing API)
and the new user I created only has access to that API.  Support will be
added in JReleaser 1.12.